requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.32.5
